Fixed how schemas are sent to registry
The schema isn't sent plain as a normal Avro schema.  Apparently, the
Confluence registry JSON is a bit different, with an encapsulating
"schema" element (see
https://docs.confluent.io/current/schema-registry/docs/api.html#post--subjects-(string-%20subject)-versions).

This schema element is a string value containing the escaped schema
string all in one line.

So, parse out the raw schema into a Json map vlaue, add the
encapsulating object, then reserialize as a string value to send to the
registry.