feat: add reqwest ClientBuilder parameter to schema registry builder
This change enables the ability to add any arbitrary change to the HTTP
client used by the registry client. Therein, we get the chance to add
TLS and more notably mTLS support (even with whichever chain we wish,
such as rustls).

This change is required as we wish to use a registry protected with
mTLS, and will therefore need special configuration from reqwest.

Signed-off-by: Mariell Hoversholm <mariell.hoversholm@paf.com>