Better errors when schema registry returns errors
Up until there were no checks of the status code of the response.
This lead to trying to parse error responses as response payload and thus failed.

The error resulting of this was hard to understand from a user PoV.
Morever, transient schema registry like 502 and 503 were considered not retryable.

What is done:
- handle status code and set it in the error content.
- make 502 and 503 retryable errors
- update tests accordingly

Fix #150

Signed-off-by: Charly Molter <charly.molter@konghq.com>